refactor: 반환 에러 수정
* not found user : 404로 수정했습니다
* not found memo : 404로 수정했습니다
